docs: added comment to targetLanguage field
Added a comment to the targetLanguage field of the Configuration, which is not used at the moment. The reason for this is that, the free-tier of the api provides 1000 requests per month, which would be exceeded after running the program once.

For that reason, the translation process is implemented, but not used.