feat: created Colorizer class
This class exposes a `colorize` method, which has two signatures. The first signature takes in a message and a color. It then adds the color prefix before the message and color-reset value at the end. This way the message is colorized without any specific variant (default -> regular).

The second signature has an additional argument, the ColorType. With the help of this argument it is possible to choose a specific variant of the request color, e.g. bold.