feat: created Console class
This class exposes a `print` method that has two signatures. The first signature takes in a sequence of message arguments. In doing so it is somewhat similar to the String.format method, since the given message can be split into multiple arguments to use variables without having to escape them.

The second signature takes in a color argument that is used to colorize the sequence of message arguments.