feat: created InputValidation class
This class exposes a method `validate` that checks whether a given string-input is valid.

The first signature takes in an input and a generic defaultValue. The method then checks whether the input is null or empty. In case its empty or null it returns the default-value. Otherwise it returns the validated input as a Integer or String based on the type of the defaultValue.

The second signature takes in an additional missingValueMessage argument, which is of type String. This message is printed when the defaultValue is used when the input is null or empty.