feat: added input-validation to requestConfiguration
The user-inputs in the `requestConfiguration` method are now validated using the previously implemented `validate` method. The reason for this is that the program crashed when no max-depth was provided. Thus, the different inputs are now validated and default-values are set.

Additionally the prompts are now colored using the `colorize` method that is exposed by the `Colorizer` class.