feat: modified Configuration to request multiple urls
When requesting the Configuration from the user, using the `Configuration`.`requestConfiguration` method, the user can now enter multiple urls. In doing so, the Configuration's url property has been renamed to urls and the type has been changed from String to String[] to store the Array of urls.

At the same time, the `crawl` method inside the Crawler has been changed to take in a given url to crawl. since it can no longer use the getUrl method from the Configuration since it has been changed to getUrls, which returns an array of urls.