feat: created `ConcurrencyHandler` class
This class is used to execute multiple Crawlers at the same time. When creating a new instance it takes in a Configuration instance. Based on the given configuration a new thread-pool is created and a callable instance for each crawl-job is created.

When the `start` method of the ConcurrencyCrawler is executed it will run all the callable-jobs that were previously created. Note, the start method will also be called automatically, when calling the `getResult` method, when it has not been called.

It also exposes a `getResults` method, which waits for each crawl-job to return its result. These results are then returned.