refactor: updated language-value of a PageInfo object
Changed the language-value that is provided to the constructor of a new PageInfo object in the Crawler from "unknown" back to an empty string (""). The reason for this is that, this way the isEmpty check simply checks whether the string is empty to determinate whether a language has been found or not.

However, to make it clear that no language was found when the value of the language property is empty, the getter for the language-field has been modified. When the language-value is equal to an empty string the getter returns "unknown". This way it is clear when a language was or was not found in a page inside the report.