fix: updated if-condition chain
Fixed the chain of the if-conditions inside of the `createReport` method in the `Printer` to log the failureReason, in case one exists, and to only print a "page is empty" when it is was retrieved and is truly empty. In other words "it is empty" should not be printed when the page is empty due to an error.