refactor: removed unnecessary property drilling
Removed unnecessary property drilling for the methods inside the `Crawler` class. The reason for this is that the arguments passed to most methods are already accessible through the crawlers private config-instance. To sum up, the arguments maxDepth and allowedDomains that were passed to most methods were removed since they accessible through the config-instance of the Crawler.