feat: created simple adapter for `Jsoup` dependency
This adapter exposes helper-function that centralize the use of the `Jsoup` dependency. In other words, if the dependency for the document-retrieval would change, this adapter would have to be changed, but the rest of the code would not change too much. In doing so the codebase is a bit more independent from the dependencies.