feat: (major) reworked structure of Crawler class
The crawler's methods were restructured in a way that the recursive-crawling process is now done inside of one method, for now.

At the same time, the retrieval of page has been restructured (`getPage`). In doing so the `JsoupAdapter` was used to make the code a bit more independent from the dependency that is used to fetch documents.

Simplified the implementations of the private helper-functions:
- `filterLinksByDomain`
- `removeDuplicateLinks`
- `cleanLinks`