feat: added addSubPage method - Page
This commit introduces a new method called `addSubPage` to the `Page` class, which takes in a Page instance as its only argument. This page instance is then added to the list of subpages.

With the help of this method an additional list that stores all the subpages before adding them to page is no longer necessary. In other words, this new method makes it possible to add the crawl-result of a subPage, directly, without temporarily storing it in a list, in the (origin / parent) page.

By adding this method the recursive-crawling method can now add each subPage to